Support a customized TTL value
Potential suggestion for supporting a customized ttl.

In some cases we may want to supply an algorithm to indicate
the cache expiry value.

For example, we may want to randomize the ttl to some degree
to avoid having a large number of keys expire at the same time.

One approach is to re-use the ttl parameter but allow it to be a number
or a custom function.